Remove static configurations from default CI builds
I didn't foresee the problem with the resources: they aren't included in
a static library, making BugTrap unable to show all of it's UI dialogs,
because they are resources.
Solution for this is shipping the resources along with the .lib so that
BugTrap user can link them to a final DLL/EXE.
Another problem is a DllMain procedure that can make problems too, if a
user already has this procedure.

I'll leave the configurations themselves, if someone wants to do this.